backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
django-extensions==3.2.3
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
Pillow==10.4.0
psycopg2-binary==2.9.9
redis==5.1.1
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
gunicorn==23.0.0
whitenoise==6.7.0
python-decouple==3.8
dj-database-url==2.2.0
djangorestframework-simplejwt==5.3.1
django-import-export==4.1.1
openpyxl==3.1.5
pandas==2.2.3
reportlab==4.2.5
python-slugify==8.0.4
django-audit-log==0.7.0
